feat: implements chacha20poly1305
